+++ app/test/integration_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/celestiaorg/celestia-app/test/util/testnode"
 	"github.com/stretchr/testify/require"
 
-	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	cosmosnet "github.com/cosmos/cosmos-sdk/testutil/network"
 
@@ -90,132 +89,6 @@ func (s *IntegrationTestSuite) WaitForBlocks(blockCount int) {
 	for i := 0; i < blockCount; i++ {
 		err := s.network.WaitForNextBlock()
 		s.Require().NoError(err)
-	}
-}
-
-func (s *IntegrationTestSuite) TestMaxBlockSize() {
-	require := s.Require()
-	assert := s.Assert()
-	val := s.network.Validators[0]
-
-	// tendermint's default tx size limit is 1Mb, so we get close to that
-	equallySized1MbTxGen := func(c client.Context) []coretypes.Tx {
-		return blobfactory.RandBlobTxsWithAccounts(
-			s.cfg.TxConfig.TxEncoder(),
-			s.kr,
-			c.GRPCClient,
-			950000,
-			1,
-			false,
-			s.cfg.ChainID,
-			s.accounts[:20],
-		)
-	}
-
-	// Tendermint's default tx size limit is 1 MiB, so we get close to that by
-	// generating transactions of size 600 KiB because 3 blobs per transaction *
-	// 200,000 bytes each = 600,000 total bytes = 600 KiB per transaction.
-	randMultiBlob1MbTxGen := func(c client.Context) []coretypes.Tx {
-		return blobfactory.RandBlobTxsWithAccounts(
-			s.cfg.TxConfig.TxEncoder(),
-			s.kr,
-			c.GRPCClient,
-			200000, // 200 KiB
-			3,
-			false,
-			s.cfg.ChainID,
-			s.accounts[20:40],
-		)
-	}
-
-	// Generate 80 randomly sized txs (max size == 100kb). Generate these
-	// transactions using some of the same accounts as the previous generator to
-	// ensure that the sequence number is being utilized correctly in blob
-	// txs
-	randoTxGen := func(c client.Context) []coretypes.Tx {
-		return blobfactory.RandBlobTxsWithAccounts(
-			s.cfg.TxConfig.TxEncoder(),
-			s.kr,
-			c.GRPCClient,
-			50000,
-			8,
-			true,
-			s.cfg.ChainID,
-			s.accounts[40:120],
-		)
-	}
-
-	type test struct {
-		name        string
-		txGenerator func(clientCtx client.Context) []coretypes.Tx
-	}
-
-	tests := []test{
-		{
-			"20 1Mb txs",
-			equallySized1MbTxGen,
-		},
-		{
-			"20 1Mb multiblob txs",
-			randMultiBlob1MbTxGen,
-		},
-		{
-			"80 random txs",
-			randoTxGen,
-		},
-	}
-	for _, tc := range tests {
-		s.Run(tc.name, func() {
-			txs := tc.txGenerator(val.ClientCtx)
-			hashes := make([]string, len(txs))
-
-			for i, tx := range txs {
-				res, err := val.ClientCtx.BroadcastTxSync(tx)
-				require.NoError(err)
-				assert.Equal(abci.CodeTypeOK, res.Code)
-				if res.Code != abci.CodeTypeOK {
-					continue
-				}
-				hashes[i] = res.TxHash
-			}
-
-			s.WaitForBlocks(8)
-
-			heights := make(map[int64]int)
-			for _, hash := range hashes {
-				resp, err := testnode.QueryTx(val.ClientCtx, hash, true)
-				require.NoError(err)
-				assert.NotNil(resp)
-				if resp == nil {
-					continue
-				}
-				require.Equal(abci.CodeTypeOK, resp.TxResult.Code, resp.TxResult.Log)
-				heights[resp.Height]++
-				// ensure that some gas was used
-				require.GreaterOrEqual(resp.TxResult.GasUsed, int64(10))
-				// require.True(resp.Proof.VerifyProof())
-			}
-
-			require.Greater(len(heights), 0)
-
-			sizes := []uint64{}
-			// check the square size
-			for height := range heights {
-				node, err := val.ClientCtx.GetNode()
-				require.NoError(err)
-				blockRes, err := node.Block(context.Background(), &height)
-				require.NoError(err)
-				size := blockRes.Block.Data.SquareSize
-
-				// perform basic checks on the size of the square
-				require.LessOrEqual(size, uint64(appconsts.MaxSquareSize))
-				require.GreaterOrEqual(size, uint64(appconsts.MinSquareSize))
-				sizes = append(sizes, size)
-			}
-			// ensure that at least one of the blocks used the max square size
-			assert.Contains(sizes, uint64(appconsts.DefaultGovMaxSquareSize))
-		})
-		require.NoError(s.network.WaitForNextBlock())
 	}
 }
